@@ -10,6 +10,7 @@ import {
   ApplicationModule,
   ApplicationRef,
   Attribute,
+  CLIENT_RENDER_MODE_FLAG,
   COMPILER_OPTIONS,
   CONTAINER_HEADER_OFFSET,
   CSP_NONCE,
@@ -33,6 +34,7 @@ import {
   DEFAULT_LOCALE_ID,
   DEFER_BLOCK_CONFIG,
   DEFER_BLOCK_DEPENDENCY_INTERCEPTOR,
+  DEHYDRATED_BLOCK_REGISTRY,
   DebugElement,
   DebugEventListener,
   DebugNode,
@@ -41,6 +43,7 @@ import {
   DeferBlockState,
   DestroyRef,
   Directive,
+  ENABLE_ROOT_COMPONENT_BOOTSTRAP,
   ENVIRONMENT_INITIALIZER,
   EffectScheduler,
   ElementRef,
@@ -48,19 +51,20 @@ import {
   EnvironmentInjector,
   ErrorHandler,
   EventEmitter,
-  ExperimentalPendingTasks,
   FactoryTarget,
   HOST_TAG_NAME,
   Host,
   HostAttributeToken,
   HostBinding,
   HostListener,
+  HydrationStatus,
   IMAGE_CONFIG,
   IMAGE_CONFIG_DEFAULTS,
   INJECTOR$1,
   INJECTOR_SCOPE,
   INTERNAL_APPLICATION_ERROR_HANDLER,
   IS_HYDRATION_DOM_REUSE_ENABLED,
+  IS_INCREMENTAL_HYDRATION_ENABLED,
   Inject,
   InjectFlags,
   Injectable,
@@ -68,12 +72,13 @@ import {
   Injector,
   Input,
   IterableDiffers,
+  JSACTION_BLOCK_ELEMENT_MAP,
   JSACTION_EVENT_CONTRACT,
   KeyValueDiffers,
   LContext,
   LOCALE_ID,
-  LifecycleHooksFeature,
   LocaleDataIndex,
+  MicrotaskEffectScheduler,
   MissingTranslationStrategy,
   ModuleWithComponentFactories,
   NG_COMP_DEF,
@@ -98,18 +103,26 @@ import {
   Output,
   OutputEmitterRef,
   PACKAGE_ROOT_URL,
+  PERFORMANCE_MARK_PREFIX,
   PLATFORM_ID,
   PLATFORM_INITIALIZER,
   PROVIDED_NG_ZONE,
   PendingTasks,
+  PendingTasksInternal,
   Pipe,
   PlatformRef,
   Query,
   QueryList,
+  R3Injector,
+  REQUEST,
+  REQUEST_CONTEXT,
+  RESPONSE_INIT,
   ReflectionCapabilities,
   Renderer2,
   RendererFactory2,
   RendererStyleFlags2,
+  ResourceImpl,
+  ResourceStatus,
   RuntimeError,
   SIGNAL,
   SSR_CONTENT_INTEGRITY_MARKER,
@@ -125,6 +138,9 @@ import {
   TemplateRef,
   Testability,
   TestabilityRegistry,
+  TimerScheduler,
+  TracingAction,
+  TracingService,
   TransferState,
   Type,
   USE_RUNTIME_DEPS_TRACKER_FOR_JIT,
@@ -133,7 +149,7 @@ import {
   ViewChild,
   ViewChildren,
   ViewContainerRef,
-  ViewEncapsulation$1,
+  ViewEncapsulation,
   ViewRef,
   ViewRef$1,
   XSS_SECURITY_URL,
@@ -143,6 +159,7 @@ import {
   _sanitizeUrl,
   afterNextRender,
   afterRender,
+  afterRenderEffect,
   allowSanitizationBypassAndThrow,
   annotateForHydration,
   asNativeElements,
@@ -171,6 +188,7 @@ import {
   createInjector,
   createNgModule,
   createNgModuleRef,
+  createOrReusePlatformInjector,
   createPlatform,
   createPlatformFactory,
   defaultIterableDiffers,
@@ -180,17 +198,22 @@ import {
   destroyPlatform,
   detectChangesInViewIfRequired,
   devModeEqual,
+  disableProfiling,
   effect,
   enableProdMode,
+  enableProfiling,
   findLocaleData,
   flushModuleScopingQueueAsMuchAsPossible,
   formatRuntimeError,
   forwardRef,
   generateStandaloneInDeclarationsError,
   getAsyncClassMetadataFn,
+  getClosestComponentName,
+  getComponentDef,
   getDebugNode,
-  getDeferBlocks,
+  getDeferBlocks$1,
   getDirectives,
+  getDocument,
   getHostElement,
   getInjectableDef,
   getLContext,
@@ -217,9 +240,13 @@ import {
   isSignal,
   isStandalone,
   isSubscribable,
+  isViewDirty,
+  linkedSignal,
   makeEnvironmentProviders,
   makeStateKey,
+  markForRefresh,
   mergeApplicationConfig,
+  microtaskEffect,
   model,
   noSideEffects,
   numberAttribute,
@@ -227,9 +254,13 @@ import {
   patchComponentDefWithScope,
   performanceMarkFeature,
   platformCore,
+  provideAppInitializer,
+  provideEnvironmentInitializer,
   provideExperimentalCheckNoChangesForDebug,
   provideExperimentalZonelessChangeDetection,
+  providePlatformInitializer,
   provideZoneChangeDetection,
+  publishExternalGlobalUtil,
   readHydrationInfo,
   reflectComponentType,
   registerLocaleData,
@@ -239,6 +270,7 @@ import {
   resetJitOptions,
   resolveComponentResources,
   resolveForwardRef,
+  resource,
   restoreComponentResolutionQueue,
   runInInjectionContext,
   setAllowDuplicateNgModuleIdsForTest,
@@ -251,6 +283,8 @@ import {
   setLocaleId,
   setTestabilityGetter,
   signal,
+  startMeasuring,
+  stopMeasuring,
   store,
   stringify,
   transitiveScopesFor,
@@ -261,10 +295,10 @@ import {
   unwrapSafeValue,
   viewChild,
   viewChildren,
-  whenStable,
   withDomHydration,
   withEventReplay,
   withI18nSupport,
+  withIncrementalHydration,
   ɵINPUT_SIGNAL_BRAND_WRITE_TYPE,
   ɵgetUnknownElementStrictMode,
   ɵgetUnknownPropertyStrictMode,
@@ -273,13 +307,13 @@ import {
   ɵsetUnknownPropertyStrictMode,
   ɵunwrapWritableSignal,
   ɵɵCopyDefinitionFeature,
+  ɵɵExternalStylesFeature,
   ɵɵHostDirectivesFeature,
   ɵɵInheritDefinitionFeature,
-  ɵɵInputTransformsFeature,
   ɵɵNgOnChangesFeature,
   ɵɵProvidersFeature,
-  ɵɵStandaloneFeature,
   ɵɵadvance,
+  ɵɵattachSourceLocations,
   ɵɵattribute,
   ɵɵattributeInterpolate1,
   ɵɵattributeInterpolate2,
@@ -308,6 +342,14 @@ import {
   ɵɵdeclareLet,
   ɵɵdefer,
   ɵɵdeferEnableTimerScheduling,
+  ɵɵdeferHydrateNever,
+  ɵɵdeferHydrateOnHover,
+  ɵɵdeferHydrateOnIdle,
+  ɵɵdeferHydrateOnImmediate,
+  ɵɵdeferHydrateOnInteraction,
+  ɵɵdeferHydrateOnTimer,
+  ɵɵdeferHydrateOnViewport,
+  ɵɵdeferHydrateWhen,
   ɵɵdeferOnHover,
   ɵɵdeferOnIdle,
   ɵɵdeferOnImmediate,
@@ -340,6 +382,7 @@ import {
   ɵɵgetComponentDepsFactory,
   ɵɵgetCurrentView,
   ɵɵgetInheritedFactory,
+  ɵɵgetReplaceMetadataURL,
   ɵɵhostProperty,
   ɵɵi18n,
   ɵɵi18nApply,
@@ -404,6 +447,7 @@ import {
   ɵɵrepeaterCreate,
   ɵɵrepeaterTrackByIdentity,
   ɵɵrepeaterTrackByIndex,
+  ɵɵreplaceMetadata,
   ɵɵresetView,
   ɵɵresolveBody,
   ɵɵresolveDocument,
@@ -461,7 +505,7 @@ import {
   ɵɵvalidateIframeAttribute,
   ɵɵviewQuery,
   ɵɵviewQuerySignal
-} from "./chunk-VKUFYADX.js";
+} from "./chunk-LKBYGJ3F.js";
 export {
   ANIMATION_MODULE_TYPE,
   APP_BOOTSTRAP_LISTENER,
@@ -498,7 +542,6 @@ export {
   EnvironmentInjector,
   ErrorHandler,
   EventEmitter,
-  ExperimentalPendingTasks,
   HOST_TAG_NAME,
   Host,
   HostAttributeToken,
@@ -528,13 +571,18 @@ export {
   PACKAGE_ROOT_URL,
   PLATFORM_ID,
   PLATFORM_INITIALIZER,
+  PendingTasks,
   Pipe,
   PlatformRef,
   Query,
   QueryList,
+  REQUEST,
+  REQUEST_CONTEXT,
+  RESPONSE_INIT,
   Renderer2,
   RendererFactory2,
   RendererStyleFlags2,
+  ResourceStatus,
   Sanitizer,
   SecurityContext,
   Self,
@@ -552,10 +600,11 @@ export {
   ViewChild,
   ViewChildren,
   ViewContainerRef,
-  ViewEncapsulation$1 as ViewEncapsulation,
+  ViewEncapsulation,
   ViewRef,
   afterNextRender,
   afterRender,
+  afterRenderEffect,
   asNativeElements,
   assertInInjectionContext,
   assertNotInReactiveContext,
@@ -585,6 +634,7 @@ export {
   isDevMode,
   isSignal,
   isStandalone,
+  linkedSignal,
   makeEnvironmentProviders,
   makeStateKey,
   mergeApplicationConfig,
@@ -592,11 +642,15 @@ export {
   numberAttribute,
   output,
   platformCore,
+  provideAppInitializer,
+  provideEnvironmentInitializer,
   provideExperimentalCheckNoChangesForDebug,
   provideExperimentalZonelessChangeDetection,
+  providePlatformInitializer,
   provideZoneChangeDetection,
   reflectComponentType,
   resolveForwardRef,
+  resource,
   runInInjectionContext,
   setTestabilityGetter,
   signal,
@@ -605,6 +659,7 @@ export {
   viewChildren,
   ALLOW_MULTIPLE_PLATFORMS as ɵALLOW_MULTIPLE_PLATFORMS,
   AfterRenderManager as ɵAfterRenderManager,
+  CLIENT_RENDER_MODE_FLAG as ɵCLIENT_RENDER_MODE_FLAG,
   CONTAINER_HEADER_OFFSET as ɵCONTAINER_HEADER_OFFSET,
   ChangeDetectionScheduler as ɵChangeDetectionScheduler,
   ChangeDetectionSchedulerImpl as ɵChangeDetectionSchedulerImpl,
@@ -613,19 +668,24 @@ export {
   DEFAULT_LOCALE_ID as ɵDEFAULT_LOCALE_ID,
   DEFER_BLOCK_CONFIG as ɵDEFER_BLOCK_CONFIG,
   DEFER_BLOCK_DEPENDENCY_INTERCEPTOR as ɵDEFER_BLOCK_DEPENDENCY_INTERCEPTOR,
+  DEHYDRATED_BLOCK_REGISTRY as ɵDEHYDRATED_BLOCK_REGISTRY,
   DeferBlockBehavior as ɵDeferBlockBehavior,
   DeferBlockState as ɵDeferBlockState,
+  ENABLE_ROOT_COMPONENT_BOOTSTRAP as ɵENABLE_ROOT_COMPONENT_BOOTSTRAP,
   EffectScheduler as ɵEffectScheduler,
+  HydrationStatus as ɵHydrationStatus,
   IMAGE_CONFIG as ɵIMAGE_CONFIG,
   IMAGE_CONFIG_DEFAULTS as ɵIMAGE_CONFIG_DEFAULTS,
   INJECTOR_SCOPE as ɵINJECTOR_SCOPE,
   ɵINPUT_SIGNAL_BRAND_WRITE_TYPE,
   INTERNAL_APPLICATION_ERROR_HANDLER as ɵINTERNAL_APPLICATION_ERROR_HANDLER,
   IS_HYDRATION_DOM_REUSE_ENABLED as ɵIS_HYDRATION_DOM_REUSE_ENABLED,
+  IS_INCREMENTAL_HYDRATION_ENABLED as ɵIS_INCREMENTAL_HYDRATION_ENABLED,
+  JSACTION_BLOCK_ELEMENT_MAP as ɵJSACTION_BLOCK_ELEMENT_MAP,
   JSACTION_EVENT_CONTRACT as ɵJSACTION_EVENT_CONTRACT,
   LContext as ɵLContext,
-  LifecycleHooksFeature as ɵLifecycleHooksFeature,
   LocaleDataIndex as ɵLocaleDataIndex,
+  MicrotaskEffectScheduler as ɵMicrotaskEffectScheduler,
   NG_COMP_DEF as ɵNG_COMP_DEF,
   NG_DIR_DEF as ɵNG_DIR_DEF,
   NG_ELEMENT_ID as ɵNG_ELEMENT_ID,
@@ -637,17 +697,23 @@ export {
   NO_CHANGE as ɵNO_CHANGE,
   NgModuleFactory as ɵNgModuleFactory,
   NoopNgZone as ɵNoopNgZone,
+  PERFORMANCE_MARK_PREFIX as ɵPERFORMANCE_MARK_PREFIX,
   PROVIDED_NG_ZONE as ɵPROVIDED_NG_ZONE,
-  PendingTasks as ɵPendingTasks,
+  PendingTasksInternal as ɵPendingTasksInternal,
+  R3Injector as ɵR3Injector,
   ReflectionCapabilities as ɵReflectionCapabilities,
   ComponentFactory as ɵRender3ComponentFactory,
   ComponentRef as ɵRender3ComponentRef,
   NgModuleRef as ɵRender3NgModuleRef,
+  ResourceImpl as ɵResourceImpl,
   RuntimeError as ɵRuntimeError,
   SIGNAL as ɵSIGNAL,
   SSR_CONTENT_INTEGRITY_MARKER as ɵSSR_CONTENT_INTEGRITY_MARKER,
   TESTABILITY as ɵTESTABILITY,
   TESTABILITY_GETTER as ɵTESTABILITY_GETTER,
+  TimerScheduler as ɵTimerScheduler,
+  TracingAction as ɵTracingAction,
+  TracingService as ɵTracingService,
   USE_RUNTIME_DEPS_TRACKER_FOR_JIT as ɵUSE_RUNTIME_DEPS_TRACKER_FOR_JIT,
   ViewRef$1 as ɵViewRef,
   XSS_SECURITY_URL as ɵXSS_SECURITY_URL,
@@ -670,19 +736,25 @@ export {
   compilePipe as ɵcompilePipe,
   convertToBitFlags as ɵconvertToBitFlags,
   createInjector as ɵcreateInjector,
+  createOrReusePlatformInjector as ɵcreateOrReusePlatformInjector,
   defaultIterableDiffers as ɵdefaultIterableDiffers,
   defaultKeyValueDiffers as ɵdefaultKeyValueDiffers,
   depsTracker as ɵdepsTracker,
   detectChangesInViewIfRequired as ɵdetectChangesInViewIfRequired,
   devModeEqual as ɵdevModeEqual,
+  disableProfiling as ɵdisableProfiling,
+  enableProfiling as ɵenableProfiling,
   findLocaleData as ɵfindLocaleData,
   flushModuleScopingQueueAsMuchAsPossible as ɵflushModuleScopingQueueAsMuchAsPossible,
   formatRuntimeError as ɵformatRuntimeError,
   generateStandaloneInDeclarationsError as ɵgenerateStandaloneInDeclarationsError,
   getAsyncClassMetadataFn as ɵgetAsyncClassMetadataFn,
+  getClosestComponentName as ɵgetClosestComponentName,
+  getComponentDef as ɵgetComponentDef,
   getDebugNode as ɵgetDebugNode,
-  getDeferBlocks as ɵgetDeferBlocks,
+  getDeferBlocks$1 as ɵgetDeferBlocks,
   getDirectives as ɵgetDirectives,
+  getDocument as ɵgetDocument,
   getHostElement as ɵgetHostElement,
   getInjectableDef as ɵgetInjectableDef,
   getLContext as ɵgetLContext,
@@ -703,9 +775,13 @@ export {
   isNgModule as ɵisNgModule,
   isPromise as ɵisPromise,
   isSubscribable as ɵisSubscribable,
+  isViewDirty as ɵisViewDirty,
+  markForRefresh as ɵmarkForRefresh,
+  microtaskEffect as ɵmicrotaskEffect,
   noSideEffects as ɵnoSideEffects,
   patchComponentDefWithScope as ɵpatchComponentDefWithScope,
   performanceMarkFeature as ɵperformanceMarkFeature,
+  publishExternalGlobalUtil as ɵpublishExternalGlobalUtil,
   readHydrationInfo as ɵreadHydrationInfo,
   registerLocaleData as ɵregisterLocaleData,
   renderDeferBlockState as ɵrenderDeferBlockState,
@@ -724,6 +800,8 @@ export {
   setLocaleId as ɵsetLocaleId,
   ɵsetUnknownElementStrictMode,
   ɵsetUnknownPropertyStrictMode,
+  startMeasuring as ɵstartMeasuring,
+  stopMeasuring as ɵstopMeasuring,
   store as ɵstore,
   stringify as ɵstringify,
   transitiveScopesFor as ɵtransitiveScopesFor,
@@ -732,19 +810,19 @@ export {
   unregisterAllLocaleData as ɵunregisterLocaleData,
   unwrapSafeValue as ɵunwrapSafeValue,
   ɵunwrapWritableSignal,
-  whenStable as ɵwhenStable,
   withDomHydration as ɵwithDomHydration,
   withEventReplay as ɵwithEventReplay,
   withI18nSupport as ɵwithI18nSupport,
+  withIncrementalHydration as ɵwithIncrementalHydration,
   ɵɵCopyDefinitionFeature,
+  ɵɵExternalStylesFeature,
   FactoryTarget as ɵɵFactoryTarget,
   ɵɵHostDirectivesFeature,
   ɵɵInheritDefinitionFeature,
-  ɵɵInputTransformsFeature,
   ɵɵNgOnChangesFeature,
   ɵɵProvidersFeature,
-  ɵɵStandaloneFeature,
   ɵɵadvance,
+  ɵɵattachSourceLocations,
   ɵɵattribute,
   ɵɵattributeInterpolate1,
   ɵɵattributeInterpolate2,
@@ -773,6 +851,14 @@ export {
   ɵɵdeclareLet,
   ɵɵdefer,
   ɵɵdeferEnableTimerScheduling,
+  ɵɵdeferHydrateNever,
+  ɵɵdeferHydrateOnHover,
+  ɵɵdeferHydrateOnIdle,
+  ɵɵdeferHydrateOnImmediate,
+  ɵɵdeferHydrateOnInteraction,
+  ɵɵdeferHydrateOnTimer,
+  ɵɵdeferHydrateOnViewport,
+  ɵɵdeferHydrateWhen,
   ɵɵdeferOnHover,
   ɵɵdeferOnIdle,
   ɵɵdeferOnImmediate,
@@ -805,6 +891,7 @@ export {
   ɵɵgetComponentDepsFactory,
   ɵɵgetCurrentView,
   ɵɵgetInheritedFactory,
+  ɵɵgetReplaceMetadataURL,
   ɵɵhostProperty,
   ɵɵi18n,
   ɵɵi18nApply,
@@ -870,6 +957,7 @@ export {
   ɵɵrepeaterCreate,
   ɵɵrepeaterTrackByIdentity,
   ɵɵrepeaterTrackByIndex,
+  ɵɵreplaceMetadata,
   ɵɵresetView,
   ɵɵresolveBody,
   ɵɵresolveDocument,
@@ -928,4 +1016,3 @@ export {
   ɵɵviewQuery,
   ɵɵviewQuerySignal
 };
-//# sourceMappingURL=@angular_core.js.map
